requirements.txt
onvif-zeep==0.2.12
requests
